Use NotFittedException.
