+++ src/pydvl/valuation/base.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import Iterable
 
+from pydvl.utils.exceptions import NotFittedException
 from pydvl.valuation.dataset import Dataset
 from pydvl.valuation.result import ValuationResult
 
@@ -28,7 +29,7 @@ class Valuation(ABC):
             The result of the valuation.
         """
         if not self.is_fitted:
-            raise RuntimeError("Valuation is not fitted")
+            raise NotFittedException(type(self))
         assert self.result is not None
 
         from copy import copy
